перенес переменные
